Move `useModel` hook to hooks directory, import in `ChatUI`
